pub_sub/solar_panel/requirements_old.txt
numpy
datetime
uuid
logging
argparse
google.auth
faker
google.cloud
